Make AccumuloRecordWriter its own class
